requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jsonschema>=4.0.0
PyYAML>=6.0
requests>=2.31.0         # For calling external APIs like searchapi.io
rich>=13.7.0     
uvicorn[standard]
fastapi[all]